+++ lib/sefaz/exception.rb
@@ -5,6 +5,9 @@ module SEFAZ
   class Error < StandardError
   end
 
+  class ValidationError < StandardError
+  end
+
   class NotImplemented < RuntimeError
   end