+++ test/sefaz/webservice/nfe/test_auditor.rb
@@ -5,7 +5,7 @@ require "test_helper"
 class SEFAZ::Webservice::NFE::TestAuditor < Minitest::Test
 
   def test_is_the_auditor_is_connected
-    xml = File.read("test/fixtures/nfe-valida-autorizada.xml")
+    xml = File.read("test/fixtures/NFe/NFe-enviarNF-Autorizado.xml")
     auditor = SEFAZ::Webservice::NFE::Auditor.new(xml)
     stat, msg = auditor.exec(60, 60)
     assert_equal stat, :ok