+++ test/sefaz/webservice/nfe/test_validator.rb
@@ -5,7 +5,7 @@ require "test_helper"
 class SEFAZ::Webservice::NFE::TestValidator < Minitest::Test
 
   def test_is_the_validador_is_connected
-    xml = File.read("test/fixtures/nfe-valida-autorizada.xml")
+    xml = File.read("test/fixtures/NFe/NFe-enviarNF-Autorizado.xml")
     validator = SEFAZ::Webservice::NFE::Validator.new(xml)
     stat, msg, err = validator.exec(60, 60)
     assert_equal stat, :ok